requirements.txt
numpy>=1.21.6
Pillow>=10.1.0
open3d>=0.15.2